backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ─────────────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# ── AI / ML ──────────────────────────────────────────────────────
openai-whisper==20240930
openai==1.58.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# ── Configuration ────────────────────────────────────────────────
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ── Utilities ────────────────────────────────────────────────────
numpy==1.26.4